Extract ProjectMetricsTab to separate file
Extracts the ProjectMetricsTab component from metrics-section.tsx into
its own file at project-tab.tsx for better code organization and
maintainability. The component continues to display KPI cards, cost burn
chart, throughput charts, model distribution donut, cycle time, and
success rate charts, all filtered by the TimeRangeSelector.
